+++ nagra/writer.py
@@ -150,12 +150,12 @@ class WriterMixin:
             if df[col].dtype == "object":
                 # bytes is not a dedicated type, we rely on the first
                 # value and hope the column type is consistent
-                if isinstance(df[col][0], bytes):
+                if isinstance(df[col].iloc[0], bytes):
                     continue
             if not is_copy:
                 df = df.copy()
                 is_copy = True
-            df[col] = df[col].astype(str)
+            df[col] = df[col].astype(str)  # Fixme, may create "None" strings in db
 
         rows = df[self.columns].values
         return self.executemany(rows)